app/requirements.txt
alembic==1.10.4
bandit==1.7.5
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mysql-connector-python==8.0.33
pbr==5.11.1
protobuf==3.20.3
Pygments==2.15.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
rich==13.4.1
smmap==5.0.0
SQLAlchemy==2.0.12
stevedore==5.1.0
typing_extensions==4.5.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.15.0
